+++ CHANGELOG.md
@@ -5,6 +5,13 @@ file is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and
 this project adheres to [Semantic
 Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Implemented `get_trusted_domain_list` method in the client API to fetch the
+  list of trusted domains from the server.
+
 ## [0.4.2] - 2024-07-31
 
 ### Added
@@ -181,6 +188,7 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 - `client::handshake` implements the application-level handshake process for the
   client after a QUIC connection is established.
 
+[Unreleased]: https://github.com/petabi/review-protocol/compare/0.4.2...main
 [0.4.2]: https://github.com/petabi/review-protocol/compare/0.4.1...0.4.2
 [0.4.1]: https://github.com/petabi/review-protocol/compare/0.4.0...0.4.1
 [0.4.0]: https://github.com/petabi/review-protocol/compare/0.3.0...0.4.0